@@ -1,6 +1,6 @@
 package com.appsflyer.support.publisher
 
-import android.support.v7.widget.RecyclerView
+import androidx.recyclerview.widget.RecyclerView
 import android.view.View
 import android.view.ViewGroup
 
@@ -24,19 +24,11 @@ class CardRecyclerAdapter<CardItem: CardBindable>() : RecyclerView.Adapter<CardV
     }
     fun get(position: Int): CardItem? = if (position<list.size) list.get(position) else null
 
-    fun getAll(): List<CardItem> = list
 
     override fun onBindViewHolder(vh: CardViewHolder, position: Int) {
         list.takeIf { vh.adapterPosition < list.size }?.get(vh.adapterPosition)?.let {
-//            vh.setAdvertise(it.adText, it.adImage)
-//            vh.setArticle(it.articleText, it.articleImage)
-//            vh.setOnClickAdListenr {
-//                with(list.get(vh.adapterPosition)) { onClick?.invoke(this) }
-//            }
-//            vh.setOnLongClickAdListener {
-//                with(list.get(vh.adapterPosition)) { onLongClick?.invoke(this)}
-//            }
             vh.setText(it.type, it.text, it.image)
+
         }
     }
 
