@@ -1,6 +1,6 @@
 package com.appsflyer.support.publisher
 
-import android.support.v7.widget.RecyclerView
+import androidx.recyclerview.widget.RecyclerView
 import android.view.View
 import android.view.View.OnLongClickListener
 
